scripts - bwmenu-location-wrapper: add new script (bitwarden rofi)
Wraps bwmenu to deal with annoying issue: if the mouse pointer location
intersects with the rofi popup, the focused window might change when the
popup exits. This happens because the focused window may not be the
window that the mouse pointer is currently at, because of keyboard-based
window focus.
This wrapper circumvents this by moving the mouse pointer to the top
left of the screen, so it definitely doesn't intersect with the rofi
popup. After bwmenu is done (i.e. cancelled or output has been typed),
the mouse pointer is moved back to the original position.